:electron: Add initial telemetry system
Add an initial pass at the new telemetry system. This adds a
command-line utility 'telemeter' that can be used to send all files or
specific streams/files/limit the number of files etc.

This also required a slight reworking of the configuration file. In the
interest of moving towards the TOML standard used elsewhere in Python
(e.g. pyproject.toml for package build specs etc), the config files are
now .toml instead of YAML (.yml) files.