:wrench: Add ip and model config parameters
Add `ip` and `model` as config parameters for each instrument, rather
than constructing the IP for each instrument from the subnet and
the final byte. This means we don't need to do any special handling
for cases when the instrument is a serial instrument, such as the
Vaisala CO2 soil probe.