:atom: Change how data files are migrated
Change how data files are migrated both on the nodes and when they
arrive at the server.

Now, each data acquisition driver just grabs a file and deposits in a
'receive' directory. The system then uses the inotify library to monitor
these directories and take an action when a new file is written to them.
they are first locally archived, then moved to the 'transmit' directory.
Files detected in the transmit directory are then telemetered down the
chain, depending on the telemetry configuration (radio or upload
server).